GP into baseTheme (consolidation).
