+++ src/custom/theme/baseTheme.tsx
@@ -1,10 +1,18 @@
-import { css } from 'styled-components'
-
 import Logo from 'assets/svg/logo.svg'
 import LogoDark from 'assets/svg/logo_white.svg'
 
 import { Colors } from 'theme/styled'
 import { colors as colorsUniswap } from '@src/theme'
+
+import { DefaultTheme, ThemeProvider as StyledComponentsThemeProvider, createGlobalStyle, css } from 'styled-components'
+import React, { useMemo } from 'react'
+
+import {
+  theme as themeUniswap,
+  FixedGlobalStyle as FixedGlobalStyleUniswap,
+  ThemedGlobalStyle as ThemedGlobalStyleUniswap
+} from '@src/theme'
+import { useIsDarkMode } from 'state/user/hooks'
 
 export { TYPE } from '@src/theme'
 export * from '@src/theme/components'
@@ -50,16 +58,23 @@ export function themeVariables(darkMode: boolean, colorsTheme: Colors) {
   return {
     body: {
       background: css`
-        background: ${colorsTheme.primary1};
+        background: radial-gradient(50% 50%, ${colorsTheme.primary1} 0%, ${colorsTheme.bg1} 100%) 0 -30vh no-repeat;
       `
     },
     logo: { src: `${darkMode ? LogoDark : Logo}`, alt: 'GP Logo', width: '24px', height: 'auto' },
+    appBody: {
+      boxShadow: `0px 0px 1px rgba(0, 0, 0, 0.01), 0px 4px 8px rgba(0, 0, 0, 0.04), 0px 16px 24px rgba(0, 0, 0, 0.04),
+  0px 24px 32px rgba(0, 0, 0, 0.01)`,
+      borderRadius: '30px',
+      border: 'none',
+      padding: '1rem'
+    },
     header: {
       border: `1px solid ${colorsTheme.border}`
     },
     buttonPrimary: {
       background: css`
-        background-image: linear-gradient(270deg, ${colorsTheme.purple} 30%, ${colorsTheme.blue1} 70%);
+        background: transparent linear-gradient(270deg, ${colorsTheme.purple} 30%, ${colorsTheme.blue1} 70%);
       `,
       fontSize: '16px',
       fontWeight: '500',
@@ -68,19 +83,60 @@ export function themeVariables(darkMode: boolean, colorsTheme: Colors) {
       boxShadow: `none`
     },
     buttonLight: {
-      background: css`
-        background: ${colorsTheme.primary5};
-      `,
-      backgroundHover: `${colorsTheme.primary4}`,
       fontSize: '16px',
       fontWeight: '500',
-      border: `none`,
-      borderHover: 'inherit',
-      borderRadius: '9px',
-      boxShadow: `none`
+      border: 'none',
+      borderHover: '1px solid transparent',
+      boxShadow: 'none',
+      backgroundHover: `${colorsTheme.primary4}`,
+      borderRadius: '9px'
+    },
+    currencyInput: {
+      background: `${colorsTheme.bg1}`,
+      border: `1px solid ${colorsTheme.bg2}`
+    },
+    buttonCurrencySelect: {
+      background: `linear-gradient(270deg, ${colorsTheme.purple} 0%, ${colorsTheme.blue1} 100%)`,
+      color: `${colorsTheme.white}`,
+      colorSelected: `${colorsTheme.text1}`,
+      boxShadow: '0px 6px 10px rgba(0, 0, 0, 0.075)'
     },
     bgLinearGradient: css`
       background-image: linear-gradient(270deg, ${colorsTheme.purple} 30%, ${colorsTheme.blue1} 70%);
-    `
+    `,
+    version: colorsTheme.green1
   }
 }
+
+export function theme(darkmode: boolean): DefaultTheme {
+  const colorsTheme = colors(darkmode)
+  return {
+    ...themeUniswap(darkmode),
+    ...colorsTheme,
+
+    // Overide Theme
+    ...themeVariables(darkmode, colorsTheme)
+  }
+}
+
+export default function ThemeProvider({ children }: { children: React.ReactNode }) {
+  const darkMode = useIsDarkMode()
+
+  const themeObject = useMemo(() => theme(darkMode), [darkMode])
+
+  return <StyledComponentsThemeProvider theme={themeObject}>{children}</StyledComponentsThemeProvider>
+}
+
+export const FixedGlobalStyle = createGlobalStyle`
+  // Uniswap default
+  ${FixedGlobalStyleUniswap}
+`
+
+export const ThemedGlobalStyle = createGlobalStyle`
+  // Uniswap default
+  ${ThemedGlobalStyleUniswap}
+
+  html {
+    background-image: ${({ theme }) => `linear-gradient(0deg, ${theme.bg1} 0%, ${theme.bg2} 100%)`};
+  }
+`
