+++ src/custom/theme/index.tsx
@@ -3,5 +3,5 @@ export { default } from './cowSwapTheme'
 export * from './cowSwapTheme'
 
 // GP theme
-// export { default } from './gnosisProtocolTheme'
-// export * from './gnosisProtocolTheme'
+// export { default } from './baseTheme'
+// export * from './baseTheme'
